Move files to allow importing as module
