+++ .gitignore
@@ -1,0 +1,4 @@
+*.pyc
+__pycache__
+.vscode
+*.swp
