+++ brpylib/brpylib.py
@@ -37,7 +37,7 @@ from datetime    import datetime
 from math        import ceil
 from os          import path as ospath
 from struct      import calcsize, pack, unpack, unpack_from
-from brMiscFxns  import openfilecheck, brmiscfxns_ver
+from .brMiscFxns  import openfilecheck, brmiscfxns_ver
 
 # Version control set/check
 brpylib_ver        = "2.0.1"
